+++ backend/src/main/java/com/inklet/blog/blog_backend/entity/User.java
@@ -12,7 +12,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "username")
     private String username;
@@ -42,7 +42,7 @@ public class User {
     }
 
     //For update
-    public User(int id, String username, String nickname, String email, String password) {
+    public User(Long id, String username, String nickname, String email, String password) {
         this.id = id;
         this.username = username;
         this.nickname = nickname;
@@ -50,11 +50,11 @@ public class User {
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
